+++ src/config/angular-elements.ts
@@ -209,6 +209,54 @@ export const STANDARD_ANGULAR_ELEMENTS: {
     originalSelector: "[ngModel]",
   },
 
+  // NgForm (template-driven forms) - requires FormsModule
+  ngForm: {
+    name: "FormsModule",
+    importPath: "@angular/forms",
+    type: "directive",
+    selectors: ["ngForm", "[ngForm]"],
+    originalSelector: "[ngForm]",
+  },
+  "[ngForm]": {
+    name: "FormsModule",
+    importPath: "@angular/forms",
+    type: "directive",
+    selectors: ["ngForm", "[ngForm]"],
+    originalSelector: "[ngForm]",
+  },
+
+  // NgModelGroup (template-driven nested forms) - requires FormsModule
+  ngModelGroup: {
+    name: "FormsModule",
+    importPath: "@angular/forms",
+    type: "directive",
+    selectors: ["ngModelGroup", "[ngModelGroup]"],
+    originalSelector: "[ngModelGroup]",
+  },
+  "[ngModelGroup]": {
+    name: "FormsModule",
+    importPath: "@angular/forms",
+    type: "directive",
+    selectors: ["ngModelGroup", "[ngModelGroup]"],
+    originalSelector: "[ngModelGroup]",
+  },
+
+  // FormGroup (reactive forms) - requires ReactiveFormsModule
+  formGroup: {
+    name: "ReactiveFormsModule",
+    importPath: "@angular/forms",
+    type: "directive",
+    selectors: ["formGroup", "[formGroup]"],
+    originalSelector: "[formGroup]",
+  },
+  "[formGroup]": {
+    name: "ReactiveFormsModule",
+    importPath: "@angular/forms",
+    type: "directive",
+    selectors: ["formGroup", "[formGroup]"],
+    originalSelector: "[formGroup]",
+  },
+
   // NgPluralCase - individual plural cases
   ngPluralCase: {
     name: "NgPluralCase",
@@ -343,5 +391,42 @@ export const STANDARD_ANGULAR_ELEMENTS: {
     type: "pipe",
     selectors: ["i18nPlural"],
     originalSelector: "i18nPlural",
+  },
+
+  // Router directives - from @angular/router
+  routerLink: {
+    name: "RouterLink",
+    importPath: "@angular/router",
+    type: "directive",
+    selectors: ["routerLink", "[routerLink]"],
+    originalSelector: "[routerLink]",
+  },
+  "[routerLink]": {
+    name: "RouterLink",
+    importPath: "@angular/router",
+    type: "directive",
+    selectors: ["routerLink", "[routerLink]"],
+    originalSelector: "[routerLink]",
+  },
+  routerLinkActive: {
+    name: "RouterLinkActive",
+    importPath: "@angular/router",
+    type: "directive",
+    selectors: ["routerLinkActive", "[routerLinkActive]"],
+    originalSelector: "[routerLinkActive]",
+  },
+  "[routerLinkActive]": {
+    name: "RouterLinkActive",
+    importPath: "@angular/router",
+    type: "directive",
+    selectors: ["routerLinkActive", "[routerLinkActive]"],
+    originalSelector: "[routerLinkActive]",
+  },
+  "router-outlet": {
+    name: "RouterOutlet",
+    importPath: "@angular/router",
+    type: "directive",
+    selectors: ["router-outlet"],
+    originalSelector: "router-outlet",
   },
 };